Refactor interface handling and update imports
Removed interface handling from the Game class and commented out unused code related to interfaces. Updated import statements across various modules to reflect the new structure where poker components are split into separate files.
