+++ streamlit_app/st_app.py
@@ -1,9 +1,9 @@
 import streamlit as st
 import random
 from core.game import StreamlitInterface
-from core.poker import (PokerGame,
-                        PokerAction,
-                        PokerHand)
+from core.poker_game import (PokerGame)
+from core.poker_hand import PokerHand
+from core.poker_action import PokerAction
 from core.poker_player import PokerPlayer, AIPokerPlayer
 from core.utils import get_players, shift_list_left
